Change definition of inline functions
Upon further reading, it may make more sense to use the extern inline definition
for inline functions. This means that these functions will only ever be used for
inlining, and they will never be compiled as functions. May cause problems if
a person of the compiler tries to use these as normal functions.